feat(microgpt-rs): Plan 077 SpectralQuant GOAT + Plan 078 variance schedules + Plan 079 ELF SDE
Plan 077: SpectralQuant codebook fitting fix
- Fixed cosine=0 roundtrip bug (LloydMax codebooks never fitted)
- v2 water-fill cosine=0.65 (+0.05 vs v1 uniform)
- Tradeoff: 10.7× compression vs TurboQuant 5.3×, but cosine 0.65 vs 0.97
- Eigenbasis calibration: d_eff=8.3, spectral_gap=1.24, 3.73ms

Plan 078: RePlaid Variance-Minimized Schedules (modelless)
- VarianceMinimizer: 10.8 ns/obs, tracks correctly (✅ ship)
- AdaptiveNoiseSchedule: infrastructure ready, needs real data
- Bandit VarianceEpsilon: partial (doesn't beat UCB1)
- SdarLearnedBeta: partial (hits bounds on synthetic data)

Plan 079: ELF Embedded Language Flows (modelless) — GOAT proved
- SDE noise: 10-22× path diversity at 3.2µs overhead
- Logit-normal schedule: 2.2× step concentration at 159ns overhead
- elf_sde promoted to default-on feature

Feature changes:
- default += elf_sde (GOAT proved)
- spectral_quant stays default-on (compression advantage)
- replaid_schedules stays off-default (partial GOAT)

+++ src/spectralquant/nonuniform_quant.rs
@@ -297,7 +297,7 @@ mod tests {
                 (0..head_dim)
                     .map(|i| {
                         let scale = if i < d_eff { 1.0 } else { 0.1 };
-                        rng.normal() as f32 * scale
+                        rng.normal() * scale
                     })
                     .collect()
             })
